+++ src/test/java/org/openstreetmap/atlas/geography/atlas/change/AbstractChangeTest.java
@@ -4,7 +4,6 @@ import org.junit.Rule;
 import org.junit.rules.ExpectedException;
 import org.openstreetmap.atlas.geography.Polygon;
 import org.openstreetmap.atlas.geography.atlas.AtlasResourceLoader;
-import org.openstreetmap.atlas.geography.atlas.change.feature.FeatureChange;
 import org.openstreetmap.atlas.geography.atlas.complete.CompleteArea;
 import org.openstreetmap.atlas.geography.atlas.complete.CompleteLine;
 import org.openstreetmap.atlas.utilities.collections.Maps;
