+++ src/test/java/org/openstreetmap/atlas/geography/atlas/change/AtlasChangeGeneratorRemoveReverseEdges.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 import java.util.Set;
 
 import org.openstreetmap.atlas.geography.atlas.Atlas;
-import org.openstreetmap.atlas.geography.atlas.change.feature.FeatureChange;
 import org.openstreetmap.atlas.geography.atlas.complete.CompleteEntity;
 import org.openstreetmap.atlas.geography.atlas.items.Edge;
 import org.openstreetmap.atlas.utilities.collections.Iterables;
