+++ src/test/java/org/openstreetmap/atlas/geography/atlas/change/ChangeMergeTest.java
@@ -4,7 +4,6 @@ import java.util.stream.IntStream;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.openstreetmap.atlas.geography.atlas.change.feature.FeatureChange;
 
 /**
  * Test the merge feature of Change class.
